Drive the lobby behavior tests through the service, not the socket API
The visibility/join/closeSlot/summaries/game-config suites all exercised
LobbyService behavior but reached it through LobbyApi handler calls,
coupling every one of them to the websocket transport that is slated to
be replaced by HTTP handlers. They now live in lobby-service.test.ts and
call the service directly; join failures assert LobbyServiceErrorCodes,
with the code -> wire-shape translation already locked down exhaustively
by the convertLobbyServiceError table test - which is all that remains in
lobby-socket-api.test.ts.
